refactor: extract order creation workflow into dedicated service
