+++ backend/api/src/routes/orderRoutes.js
@@ -34,6 +34,7 @@ import {
   ESCROW_MATIC_PER_PAISA,
 } from '../services/escrow.js';
 import { BidAcceptanceService, DomainError } from '../services/order/bidAcceptanceService.js';
+import { createOrder } from '../services/order/orderCreationService.js';
 import {
   sendDeliveryOtpNotification,
   storeDeliveryOtp,
@@ -41,7 +42,7 @@ import {
   verifyDeliveryOtp,
   expireDeliveryOtps
 } from '../services/notificationService.js';
-import { predictDemand, predictPrice } from '../services/ml.js';
+import { predictDemand } from '../services/ml.js';
 import { requireIdempotency } from '../middleware/idempotency.js';
 import { acquireLock, releaseLock } from '../lib/redisLock.js';
 import logger from '../middleware/logger.js';
@@ -205,167 +206,23 @@ const changeDropLimiter = rateLimit({
   message: { error: 'Too many drop change requests. Please try again later.' },
 });
 
-/**
- * Helper to generate order display IDs like #FF20260521
- */
-function generateOrderDisplayId() {
-  const prefix = '#FF';
-  const now = new Date();
-  const dateStr = now.toISOString().slice(0, 10).replace(/-/g, ''); // YYYYMMDD
-  const random = crypto.randomInt(100000, 999999).toString(); // 6 random digits using CSPRNG
-  return `${prefix}${dateStr}${random}`;
-}
-
 // ============================================================================
 // 1. CREATE AN ORDER (CUSTOMER)
 // ============================================================================
 router.post('/', authenticate, userLimiter, requireRole(['customer']), validateBody(createOrderSchema), async (req, res) => {
-  const {
-    pickup_address, pickup_lat, pickup_lng,
-    drop_address, drop_lat, drop_lng,
-    pickup_date, pickup_time,
-    goods_type, weight_tonnes, length_ft, width_ft, height_ft,
-    is_stackable, is_fragile, special_requirements,
-    payment_method_id, upi_id
-  } = req.body;
-
-  if (!pickup_address || pickup_lat == null || pickup_lng == null || !drop_address || drop_lat == null || drop_lng == null || !goods_type || weight_tonnes == null) {
-    return res.status(400).json({ error: 'Missing required routing or cargo specification fields.' });
-  }
-
-  let pricing;
   try {
-    const routeEstimate = await getRouteEstimate({
-      pickupLat: Number(pickup_lat),
-      pickupLng: Number(pickup_lng),
-      dropLat: Number(drop_lat),
-      dropLng: Number(drop_lng),
+    const result = await createOrder({
+      orderData: req.body,
+      userId: req.user.id,
+      user: req.user,
     });
-    pricing = computeOrderPricing({
-      pickupLat:  Number(pickup_lat),
-      pickupLng:  Number(pickup_lng),
-      dropLat:    Number(drop_lat),
-      dropLng:    Number(drop_lng),
-      weightTonnes: Number(weight_tonnes),
-      roadDistanceKm: routeEstimate?.distanceKm,
-      isFragile:   Boolean(is_fragile),
-      isStackable: Boolean(is_stackable),
-    });
-  } catch (pricingErr) {
-    logger.error('Pricing computation error:', pricingErr.message);
-    return res.status(400).json({
-      error: 'Unable to compute freight pricing for the given route/cargo.',
-      details: pricingErr.message,
-    });
-  }
-
-  let estimatedPrice = null;
-  try {
-    const mlResult = await predictPrice({
-      distanceKm: pricing.distanceKm,
-      cargoWeightKg: Number(weight_tonnes) * 1000,
-      routeOrigin: pickup_address,
-      routeDestination: drop_address,
-    });
-    if (!mlResult || typeof mlResult.estimated_price !== 'number' || mlResult.estimated_price <= 0) {
-      throw new Error(`Invalid or non-positive price prediction: ${JSON.stringify(mlResult)}`);
-    }
-    estimatedPrice = Math.round(mlResult.estimated_price * 100);
-  } catch (mlErr) {
-    logger.warn({ err: mlErr.message }, 'Price prediction unavailable, falling back to base pricing');
-  }
-
-  const MAX_ID_RETRIES = 3;
-  let order = null;
-  let orderErr = null;
-  let orderDisplayId = null;
-
-  try {
-    for (let attempt = 0; attempt < MAX_ID_RETRIES; attempt++) {
-      orderDisplayId = generateOrderDisplayId();
-      const result = await supabase
-        .from('orders')
-        .insert({
-          order_display_id: orderDisplayId,
-          customer_id: req.user.id,
-          status: 'pending',
-          pickup_address, pickup_lat, pickup_lng,
-          drop_address, drop_lat, drop_lng,
-          pickup_date, pickup_time,
-          goods_type, weight_tonnes, length_ft, width_ft, height_ft,
-          is_stackable, is_fragile, special_requirements,
-          base_freight: pricing.baseFreight,
-          toll_estimate: pricing.tollEstimate,
-          platform_fee: pricing.platformFee,
-          total_amount: pricing.totalAmount,
-          estimated_price: estimatedPrice,
-          payment_method_id, upi_id
-        })
-        .select('id, order_display_id, status, created_at')
-        .single();
-
-      order = result.data;
-      orderErr = result.error;
-
-      if (!orderErr || orderErr.code !== '23505') break;
-      logger.warn(`[Orders] display ID collision on ${orderDisplayId}, retrying (attempt ${attempt + 1}/${MAX_ID_RETRIES})`);
-    }
-
-    if (orderErr) {
-      logger.error('Order Insertion Error:', orderErr.message);
-      return res.status(500).json({ error: 'Failed to create order record.', details: orderErr.message });
-    }
-
-    const milestones = [
-      { order_display_id: orderDisplayId, milestone: 'Order Placed', milestone_time: new Date().toISOString(), completed: true, sort_order: 10 },
-      { order_display_id: orderDisplayId, milestone: 'Truck Assigned', milestone_time: null, completed: false, sort_order: 20 },
-      { order_display_id: orderDisplayId, milestone: 'En Route to Pickup', milestone_time: null, completed: false, sort_order: 30 },
-      { order_display_id: orderDisplayId, milestone: 'Arrived at Pickup', milestone_time: null, completed: false, sort_order: 35 },
-      { order_display_id: orderDisplayId, milestone: 'Goods Loaded', milestone_time: null, completed: false, sort_order: 40 },
-      { order_display_id: orderDisplayId, milestone: 'In Transit', milestone_time: null, completed: false, sort_order: 50 },
-      { order_display_id: orderDisplayId, milestone: 'Arriving', milestone_time: null, completed: false, sort_order: 55 },
-      { order_display_id: orderDisplayId, milestone: 'Delivered', milestone_time: null, completed: false, sort_order: 60 }
-    ];
-
-    const { error: timelineErr } = await supabase.from('order_timeline').insert(milestones);
-
-    if (timelineErr) {
-      logger.error('Timeline Insertion Error:', timelineErr.message);
-      await supabase.from('orders').delete().eq('id', order.id);
-      return res.status(500).json({ error: 'Failed to create order timeline.', details: timelineErr.message });
-    }
-
-    const { error: offerErr } = await supabase
-      .from('load_offers')
-      .insert({
-        order_display_id: orderDisplayId,
-        customer_id: req.user.id,
-        customer_name: req.user.fullName || 'Customer',
-        route_label: `${pickup_address.split(',')[0]} → ${drop_address.split(',')[0]}`,
-        route_subtitle: `${weight_tonnes} tonnes • ${goods_type}`,
-        pickup_address, pickup_lat, pickup_lng,
-        drop_address, drop_lat, drop_lng,
-        goods_type,
-        weight: `${weight_tonnes} tonnes`,
-        freight_value: pricing.totalAmount,
-        fuel_cost: pricing.fuelCost,
-        toll_cost: pricing.tollEstimate,
-        net_profit: pricing.netProfit,
-        extra_distance_km: pricing.distanceKm,
-        status: 'available'
-      });
-
-    if (offerErr) {
-      logger.error('Load Offer Insertion Error:', offerErr.message);
-      await supabase.from('order_timeline').delete().eq('order_display_id', orderDisplayId);
-      await supabase.from('orders').delete().eq('id', order.id);
-      return res.status(500).json({ error: 'Failed to create load offer.', details: offerErr.message });
-    }
-
-    res.status(201).json({ message: 'Order created successfully and broadcasted to loads board.', order });
+    return res.status(201).json(result);
   } catch (err) {
+    if (err instanceof DomainError) {
+      return res.status(err.status).json(err.payload);
+    }
     logger.error('Order creation exception:', err.message);
-    res.status(500).json({ error: 'Internal Server Error.' });
+    return res.status(500).json({ error: 'Internal Server Error.' });
   }
 });
